+++ api/api_test.go
@@ -60,8 +60,8 @@ func (s *successGetMetrics) GetMetrics(deployment *appsv1.Deployment) (*metric.R
 		DeploymentName: deployment.Name,
 		Metrics: []*metric.Metric{
 			&metric.Metric{
-				Value: "SUCCESS",
-				Pod:   "SUCCESS_POD",
+				Value:    "SUCCESS",
+				Resource: "SUCCESS_POD",
 			},
 		},
 		Deployment: deployment,
@@ -141,7 +141,7 @@ func TestAPI(t *testing.T) {
 		},
 		{
 			"Get metrics success metric gathering",
-			"{\"deployment\":\"test\",\"metrics\":[{\"pod\":\"SUCCESS_POD\",\"value\":\"SUCCESS\"}]}",
+			"{\"deployment\":\"test\",\"metrics\":[{\"resource\":\"SUCCESS_POD\",\"value\":\"SUCCESS\"}]}",
 			http.StatusOK,
 			"GET",
 			"/metrics",
@@ -167,7 +167,7 @@ func TestAPI(t *testing.T) {
 		},
 		{
 			"Get metrics success metric gathering two deployment same namespace",
-			"{\"deployment\":\"target\",\"metrics\":[{\"pod\":\"SUCCESS_POD\",\"value\":\"SUCCESS\"}]}",
+			"{\"deployment\":\"target\",\"metrics\":[{\"resource\":\"SUCCESS_POD\",\"value\":\"SUCCESS\"}]}",
 			http.StatusOK,
 			"GET",
 			"/metrics",
@@ -200,7 +200,7 @@ func TestAPI(t *testing.T) {
 		},
 		{
 			"Get metrics success metric gathering two deployment different namespaces",
-			"{\"deployment\":\"test\",\"metrics\":[{\"pod\":\"SUCCESS_POD\",\"value\":\"SUCCESS\"}]}",
+			"{\"deployment\":\"test\",\"metrics\":[{\"resource\":\"SUCCESS_POD\",\"value\":\"SUCCESS\"}]}",
 			http.StatusOK,
 			"GET",
 			"/metrics",